OAI: Refactor embeddings
Move files and rewrite routes to adhere to Tabby's code style.
